+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.17.1 (May 25, 2021)
+
+- **[Fix]**: Fixed @opentok/types (merged from [PR505](https://github.com/opentok/opentok-react-native/pull/505))
+- **[Fix]**: Fixed double listeners [#271](https://github.com/opentok/opentok-react-native/issues/271) (merged from [PR307](https://github.com/opentok/opentok-react-native/pull/307))
+
+
+# 0.17.0 (Apr 29, 2021)
+
+- **[Feature]**: Updated to react-native 0.64.0
+- **[Fix]**: Updated npm dependencies
+- **[Fix]**: Fixed react-native 0.64.0 `nativeEvents.listeners()` crash [PR493](https://github.com/opentok/opentok-react-native/pull/493)
+
+# 0.16.0 (Apr 19, 2021)
+
+- **[Feature]**: Added support for `stereo` output on `iOS` and `Android` using `CustomAudioDevice`
+
 # 0.15.0 (Jan 26, 2021)
 
 - **[Feature]**: OTSubscriber: added [preferredResolution](https://tokbox.com/developer/sdks/js/reference/Subscriber.html#setPreferredResolution) and [preferredFrameRate](https://tokbox.com/developer/sdks/js/reference/Subscriber.html#setPreferredFrameRate) properties.